get_inline_instances patch: conditionally return generator or list
django-nested-admin overrides get_inline_instances and turns it into
a generator, which allows the surreptitious insertion of nested inlines
in the course of iterating on a parent inline. However, regular django
admin returns a list. Using an iterator in the standard django admin
causes issues because of calls to generator-unfriendly functions such
as "zip" (pre python 3.0). This fix makes sure that our monkeypatch
of ModelAdmin.get_inline_instances() returns either a generator or
a list, depending on the type returned from the super.