Don't make GenericForeignFileField editable=True for django < 1.10
This partially reverts the changes in eade93dc. The way generic plus
currently works, the model has two fields with the same name, one of
which is a FileField (which is the concrete field) and the other that is
a GenericRelation (which is a virtual field). Before Django 1.10
django-generic-plus used the fact that
django.forms.models.fields_for_model() skips fields with
editable=False, so it would only call formfield on virtual field. But in
Django 1.10+ having a field on a ModelForm that has editable=True throws
an exception. So for Django 1.10 we set editable=True and rely on the
fact that the order in which fields are iterated over in
fields_for_models() always has the virtual field evaluated after the
concrete model, so when the OrderedDict is created the formfield will be
from the virtual field.

This is obviously very fragile, which is why out of caution I'm only
relying on this for Django 1.10, and otherwise using the editable=False
trick, which seems less fragile. But in a future iteration of
django-generic-plus the concrete field and the virtual field will have
different field names, so that we can get rid of this hacky
implementation.